[ADD] GET Model Detail information by md_id parameter

See merge request s04-ai-image-sub3/s04p23b101!86

Former-commit-id: c0db99e155e90c37a416ce85dde4322d2356987c [formerly 27fe19d9c2bc7b66ae448d30efa894ab416460ae] [formerly 0c8b6e52028d9bdd1e0d51836514817863c83ae7 [formerly 54907d232981e490702876d7c6da85fa5ae6e268]]
Former-commit-id: d5ddbe916ce3820ceb3c83493ace317d0b83b8f1 [formerly 50f4d5c18146981093cf009ea627eba8e693c764]
Former-commit-id: b1574bd140abebacedf5611e99b4912efec09606

+++ backend/src/main/java/com/komencar/backend/repository/ModelDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface ModelDetailRepository extends JpaRepository<Model_Detail, Long> {
+public interface ModelDetailRepository extends JpaRepository<Model_Detail, Integer> {
 }
